pyproject.toml
[project]
name = "langgraph-youtube-pipeline"
version = "0.1.0"
description = "Automated Weekly YouTube Content Pipeline using LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai",
    "moviepy>=1.0.3",
    "openai>=1.12.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"